refactor: 단일 책임 원칙에 따른 서비스 계층 분리
